+++ src/pages/Jungwoo/Login/Login.js
@@ -2,39 +2,10 @@
 
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import BoxLogin from '../../../Components/Login/BoxLogin.js';
+import BoxPw from '../../../Components/Login/BoxPw.js';
+import Footer from '../../../Components/Login/Footer.js';
 import './Login.scss';
-
-class BoxLogin extends React.Component {
-  render(props) {
-    return (
-      <div className="boxLogin">
-        <input
-          type="text"
-          id="box"
-          className="boxProperty"
-          placeholder="전화번호, 사용자 이름 또는 이메일"
-          onChange={this.props.handleIdInput}
-        />
-      </div>
-    );
-  }
-}
-
-class BoxPw extends React.Component {
-  render(props) {
-    return (
-      <div className="boxPw">
-        <input
-          type="text"
-          id="box2"
-          className="boxProperty"
-          placeholder="비밀번호"
-          onChange={this.props.handlePwInput}
-        />
-      </div>
-    );
-  }
-}
 
 class LoginJungwoo extends React.Component {
   constructor(props) {
@@ -48,8 +19,40 @@ class LoginJungwoo extends React.Component {
   goToMain = () => {
     this.state.idInputValue.indexOf('@') !== -1 &&
     this.state.pwInputValue.length > 5
-      ? this.props.history.push('/main-jungwoo')
-      : alert('아이디와 패스워드를 바르게 입력해 주세요.');
+      ? fetch('http://192.168.29.61:8000/users/login', {
+          method: 'POST',
+          body: JSON.stringify({
+            email: this.state.idInputValue,
+            password: this.state.pwInputValue,
+          }),
+        })
+          .then(response => response.json())
+          .then(response => {
+            if (response.token) {
+              localStorage.setItem('token', response.token);
+              this.props.history.push('/main-jungwoo');
+            } else {
+              alert('아이디와 비밀번호를 다시 한 번 확인해 주세요.');
+            }
+          })
+      : alert('아이디와 비밀번호를 정확히 입력해 주세요.');
+  };
+
+  signUp = () => {
+    this.state.idInputValue.indexOf('@') !== -1 &&
+    this.state.pwInputValue.length > 5
+      ? fetch('http://192.168.29.61:8000/users/signup', {
+          method: 'POST',
+          body: JSON.stringify({
+            email: this.state.idInputValue,
+            password: this.state.pwInputValue,
+          }),
+        })
+          .then(response => response.json())
+          .then(result => {
+            console.log('결과: ', result);
+          })
+      : alert('아이디와 비밀번호를 정확히 입력해 주세요.');
   };
 
   handleIdInput = event => {
@@ -77,7 +80,7 @@ class LoginJungwoo extends React.Component {
               <div className="mainBoxContainer">
                 <h1>Instagram</h1>
                 <div className="loginBox">
-                  <form>
+                  <div>
                     <BoxLogin handleIdInput={this.handleIdInput} />
                     <BoxPw handlePwInput={this.handlePwInput} />
 
@@ -106,42 +109,18 @@ class LoginJungwoo extends React.Component {
                       </h6>
                     </div>
                     <a className="rememberPw">비밀번호를 잊으셨나요?</a>
-                  </form>
+                  </div>
                 </div>
               </div>
 
               <div className="signUp">
                 <p>
-                  계정이 없으신가요? <a>가입하기</a>
+                  계정이 없으신가요? <a onClick={this.signUp}>가입하기</a>
                 </p>
               </div>
             </div>
           </div>
-
-          <footer className="footerContainer">
-            <div className="footerBox">
-              <div>
-                <ul>
-                  <li>소개</li>
-                  <li>블로그</li>
-                  <li>채용 정보</li>
-                  <li>도움말</li>
-                  <li>API</li>
-                  <li>개인정보처리방침</li>
-                  <li>약관</li>
-                  <li>인기 계정</li>
-                  <li>해시 태그</li>
-                  <li>위치</li>
-                </ul>
-              </div>
-              <div>
-                <ul>
-                  <li>한국어</li>
-                  <li>© 2021 Instagram from Facebook</li>
-                </ul>
-              </div>
-            </div>
-          </footer>
+          <Footer />
         </div>
       </>
     );
