+++ src/pages/Jungwoo/Login/Login.scss
@@ -192,7 +192,7 @@ body {
   height: 20px;
   width: 500px !important;
   margin-top: 50px;
-  margin-left: -210px;
+  margin-left: -280px;
   li {
     @extend %footerBoxLi;
     padding: 0 3px;
